+++ lib/models/available_periodic_reminder_items.dart
@@ -6,15 +6,13 @@ class AvailablePeriodicReminderItems {
       id: 'donate',
       title: 'periodic_reminder_donate_title',
       body: 'periodic_reminder_donate_body',
-      button: 'jail_dialog_button',
-      buttonFunction: () => print('hi'),
+      button: 'periodic_reminder_backup_button',
     ),
     'backup': PeriodicReminderItem(
       id: 'backup',
       title: 'periodic_reminder_backup_title',
       body: 'periodic_reminder_backup_body',
-      button: 'periodic_reminder_backup_button',
-      buttonFunction: () => print('hi'),
+      button: 'jail_dialog_button',
     )
   };
 }