+++ lib/models/periodic_reminder_item.dart
@@ -3,13 +3,11 @@ class PeriodicReminderItem {
   final String title;
   final String body;
   final String button;
-  final Function buttonFunction;
 
   PeriodicReminderItem({
     required this.id,
     required this.title,
     required this.body,
     required this.button,
-    required this.buttonFunction,
   });
 }