+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### **0.8.7** (2022-03-xx)
 * Periodic backup reminders
-* Add currency: AUD
+* Add language: Swedish
+* Add currency: AUD, SEK
 
 ### **0.8.6** (2022-03-09)
 * Add language: Filipino